gfxstream: delete HealthMonitor and MetricsLogger
So these were added for B*, but they ended up forking anyways
and nobody really uses this.  The plan for gfxstream is to
test on Lavapipe, so we should be able to gurantee health
via the testing.

In addition, it it's probably possible to create a simpler
and more portable versions if the need arises.  For example,
now with unified logging, the gfxstream logger can possibly
emit a json file instead, which the recieving end can parse.

However, this all immaterial, since there is no upstream user
of these APIs, so remove it to simplify the code..

BUG=406071091
TEST=CI/CD

Change-Id: Ie2e04592a600cae3715c1b540cc93bbb4a1cbdee